# app/src/generated/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {},
	"info": {
		"title": "test",
		"version": "1.0.0",
		"description": "description",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/HealthCheck": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/node/{nodeId}/height": {
			"get": {
				"operationId": "GetHeight",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "nodeId",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
		"/node/{nodeId}/root": {
			"get": {
				"operationId": "GetRoot",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "nodeId",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
		"/node/{nodeId}/children": {
			"get": {
				"operationId": "GetChildren",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"items": {
								"type": "number",
								"format": "double"
							},
							"type": "array"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "nodeId",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
		"/node/{nodeId}/parent/{parentId}": {
			"post": {
				"operationId": "ChangeParent",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "nodeId",
						"required": true,
						"format": "double",
						"type": "number"
					},
					{
						"in": "path",
						"name": "parentId",
						"required": true,
						"format": "double",
						"type": "number"
					}
				]
			}
		},
		"/node": {
			"post": {
				"operationId": "NewNode",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "number",
							"format": "double"
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {},
	"host": "localhost:3010",
	"schemes": [
		"http"
	]
}